requirements.txt
# 豆包AI SDK - 尝试不同的包名
# 主要包名
volcengine-ark>=1.0.0
# 备选包名
# volcengine-sdk>=1.0.0
# volcengine>=1.0.0

# 环境变量管理
python-dotenv>=1.0.0

# HTTP请求库
requests>=2.31.0

# 图像处理
Pillow>=10.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# Word文档生成
python-docx>=0.8.11

# PDF生成
reportlab>=4.0.0
# 备选PDF库
# weasyprint>=60.0
# pdfkit>=1.0.0

# Excel导出
openpyxl>=3.1.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
